requirements.txt
# Core ML + NLP
transformers>=4.43.0
torch>=2.3.0

# Data + Dashboard
pandas>=2.2.2
plotly>=5.24.0
dash>=2.18.1

# Streaming (optional)
tweepy>=4.14.0

# Utilities
python-dotenv>=1.0.1
